Meet scheme utils in separate files. Doc / naming updates.
